Switch sum plugin to use doubles instead of BigDecimal.

+++ aurora4spark-parent/aurora4spark-sql-plugin/src/test/scala/com/nec/spark/agile/SummingSparkPlanSpec.scala
@@ -17,7 +17,7 @@ final class SummingSparkPlanSpec extends AnyFreeSpec {
       import sparkSession.implicits._
       val result =
         SummingSparkPlan
-          .summingRdd(Seq[BigDecimal](1, 2, 2).toDS().rdd, ScalaSummer)
+          .summingRdd(Seq[Double](1, 2, 2).toDS().rdd, ScalaSummer)
           .collect()
           .head
       assert(result == BigDecimal(5))
